@ptmcg 2020-03-24
Cursor testing looks okay - looking forward to seeing other optimizations.